pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "prompts"
version = "0.0.0-9001"
description = "A package for notetaking inspired by the zettelkasten method"
readme = "README.md"
dependencies = ['typer', 'python-frontmatter', 'Jinja2>=3.1', 'shellingham>=1.5']
[project.scripts]
qp = "qp.cli:app"

[tool.setuptools.package-data]
"qp.templates" = ["*.md.jinja"]
